fix(upload): Fix progress bar calculation precision
- Cast current and total to float64 before division to prevent integer truncation
- Ensures accurate progress bar width calculation for file uploads
- Improves visual accuracy of upload progress indicator

+++ datadropcli/cmd/upload.go
@@ -247,7 +247,7 @@ func formatSpeed(bytesPerSec float64) string {
 
 func printProgressBar(current, total int64, pt *progressTracker, suffix string) {
 	percent := float64(current) / float64(total) * 100
-	filled := int(float64(progressBarWidth) * current / total)
+	filled := int(float64(progressBarWidth) * float64(current) / float64(total))
 	
 	bar := strings.Repeat("█", filled) + strings.Repeat("░", progressBarWidth-filled)
